+++ CHANGELOG.md
@@ -31,6 +31,7 @@
 * Add `Projects#delete_projects` method
 * Add `Projects#report_template` method
 * Add `Projects#update_projects_settings` method
+* Add `Portal#delete_portal` method
 * Add `rspec-retry` support
 * Add `ruby-3.1` to CI
 * Run linters in CI before actual tests